refactor(cli/migrate): camelCase filenames according to codebase conventions
